Move aggregate_allocations out of reporters/tui.py
This is also used by the `summary` reporter, which is a CLI report and
not a TUI report. Moving this breaks an import time dependency from
`memray summary` on `textual`.
